feat: Namepsace Change and Abstracted Connector
