+++ src/Places/GooglePlaces.php
@@ -1,103 +1,26 @@
 <?php
 
-namespace SKAgarwal\GoogleApi\GoogleMaps;
+namespace SKAgarwal\GoogleApi\Places;
 
 use Saloon\Contracts\Authenticator;
 use Saloon\Http\Auth\QueryAuthenticator;
-use Saloon\Http\Connector;
 use Saloon\Http\Response;
-use SKAgarwal\GoogleApi\Exceptions\GooglePlacesApiException;
-use SKAgarwal\GoogleApi\GoogleMaps\Requests\FindPlace;
-use SKAgarwal\GoogleApi\GoogleMaps\Requests\NearbySearch;
-use SKAgarwal\GoogleApi\GoogleMaps\Requests\PlaceAutocomplete;
-use SKAgarwal\GoogleApi\GoogleMaps\Requests\PlaceDetails;
-use SKAgarwal\GoogleApi\GoogleMaps\Requests\QueryAutocomplete;
-use SKAgarwal\GoogleApi\GoogleMaps\Requests\TextSearch;
+use SKAgarwal\GoogleApi\Connector;
+use SKAgarwal\GoogleApi\Places\Requests\FindPlace;
+use SKAgarwal\GoogleApi\Places\Requests\NearbySearch;
+use SKAgarwal\GoogleApi\Places\Requests\PlaceAutocomplete;
+use SKAgarwal\GoogleApi\Places\Requests\PlaceDetails;
+use SKAgarwal\GoogleApi\Places\Requests\QueryAutocomplete;
+use SKAgarwal\GoogleApi\Places\Requests\TextSearch;
 
-class GoogleMaps extends Connector
+class GooglePlaces extends Connector
 {
-    /**
-     * @var string|null
-     */
-    private ?string $key;
-
-    /**
-     * @var bool
-     */
-    private bool $verifySSL = true;
-
-    /**
-     * @param  string|null  $key
-     * @param  bool|null  $verifySSL
-     *
-     * @throws \SKAgarwal\GoogleApi\Exceptions\GooglePlacesApiException
-     */
-    public function __construct(?string $key = null, ?bool $verifySSL = null)
-    {
-        $errorMessage = 'Google Places API KEY is missing.';
-
-        if (function_exists('config')) {
-            $key = $key ?? config('google.places.key');
-            $verifySSL = $verifySSL ?? config('google.places.verify_ssl');
-            $errorMessage = 'Google Places API KEY is not set in google config file.';
-        }
-
-        if (!$key) {
-            throw new GooglePlacesApiException($errorMessage);
-        }
-
-        $this->key = $key;
-        $this->verifySSL = $verifySSL ?? true;
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getKey(): ?string
-    {
-        return $this->key;
-    }
-
-    /**
-     * @param  string  $key
-     *
-     * @return $this
-     */
-    public function setKey(string $key): static
-    {
-        $this->key = $key;
-
-        return $this;
-    }
-
-    /**
-     * @param  bool  $verifySSL
-     *
-     * @return static
-     */
-    public function verifySSL(bool $verifySSL = true): static
-    {
-        $this->verifySSL = $verifySSL;
-
-        return $this;
-    }
-
     /**
      * @return string
      */
     public function resolveBaseUrl(): string
     {
-        return Endpoint::BASE->value;
-    }
-
-    /**
-     * @return bool[]
-     */
-    protected function defaultConfig(): array
-    {
-        return [
-            'verify' => $this->verifySSL,
-        ];
+        return Endpoint::BASE_URL->value;
     }
 
     /**
